Move GenericBaseDatastore to datastore from datastores.
