+++ python/lsst/daf/butler/datastore/generic_base.py
